# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LicitAIM - Sistema de Monitoramento Inteligente de Licitações Públicas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alertas:
    get:
      operationId: listAlertas
      tags: [alertas]
      summary: Listar alertas do usuário
      parameters:
        - name: lido
          in: query
          schema:
            type: boolean
        - name: tipo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertaPage"
...
  /alertas/{id}/ler:
    post:
      operationId: marcarAlertaLido
      tags: [alertas]
      summary: Marcar alerta como lido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marcado como lido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alerta"
...
  /equipe:
    get:
      operationId: getEquipe
      tags: [equipe]
      summary: Listar membros da equipe
      responses:
        "200":
          description: Membros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembroEquipe"
    post:
      operationId: convidarMembro
      tags: [equipe]
      summary: Convidar membro para a equipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConviteInput"
      responses:
        "201":
          description: Convite enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembroEquipe"
  /equipe/{id}:
    patch:
      operationId: updateMembro
      tags: [equipe]
      summary: Atualizar permissões do membro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembroUpdate"
      responses:
        "200":
          description: Membro atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembroEquipe"
    delete:
      operationId: removerMembro
      tags: [equipe]
      summary: Remover membro da equipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ── Oportunidades ────────────────────────────────────────────────────────────
  /oportunidades:
    get:
      operationId: listOportunidades
      tags: [oportunidades]
      summary: Listar oportunidades no pipeline
      parameters:
        - name: estagio
          in: query
          schema:
            type: string
            enum: [identificada, qualificada, proposta, disputa, ganhou, perdeu]
        - name: responsavelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de oportunidades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Oportunidade"
    post:
      operationId: createOportunidade
      tags: [oportunidades]
      summary: Adicionar oportunidade ao pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OportunidadeInput"
      responses:
        "201":
          description: Oportunidade criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oportunidade"
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── User ───────────────────────────────────────────────────────────────────
...
    Alerta:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [nova_licitacao, prazo_vencendo, situacao_alterada, nova_disputa, preco_referencia]
        titulo:
          type: string
        descricao:
          type: string
        lido:
          type: boolean
        licitacaoId:
          type: ["string", "null"]
        licitacaoObjeto:
          type: ["string", "null"]
        monitoramentoNome:
          type: ["string", "null"]
        criadoEm:
          type: string
      required:
        - id
        - tipo
        - titulo
        - descricao
        - lido
        - criadoEm
    AlertaPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Alerta"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
        totalNaoLidos:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
        - totalNaoLidos
...
    MembroEquipe:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        papel:
          type: string
          enum: [admin, editor, visualizador]
        status:
          type: string
          enum: [ativo, pendente, inativo]
        avatarUrl:
          type: ["string", "null"]
        criadoEm:
          type: string
      required:
        - id
        - nome
        - email
        - papel
        - status
        - criadoEm
    ConviteInput:
      type: object
      properties:
        email:
          type: string
        papel:
          type: string
          enum: [admin, editor, visualizador]
      required:
        - email
        - papel
    MembroUpdate:
      type: object
      properties:
        papel:
          type: string
          enum: [admin, editor, visualizador]
        status:
          type: string
          enum: [ativo, inativo]

    # ── Oportunidade ──────────────────────────────────────────────────────────
    Oportunidade:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        estagio:
          type: string
          enum: [identificada, qualificada, proposta, disputa, ganhou, perdeu]
        valorEstimado:
          type: ["number", "null"]
        probabilidade:
          type: ["integer", "null"]
        licitacaoId:
          type: ["string", "null"]
        licitacaoObjeto:
          type: ["string", "null"]
        responsavelNome:
          type: ["string", "null"]
        responsavelId:
          type: ["integer", "null"]
        prazo:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        criadoEm:
          type: string
        atualizadoEm:
          type: string
      required:
        - id
        - titulo
        - estagio
        - tags
        - criadoEm
        - atualizadoEm
    OportunidadeInput:
      type: object
      properties:
        titulo:
          type: string
          minLength: 1
        estagio:
          type: string
          enum: [identificada, qualificada, proposta, disputa, ganhou, perdeu]
        valorEstimado:
          type: number
        probabilidade:
          type: integer
          minimum: 0
          maximum: 100
        licitacaoId:
          type: string
        responsavelId:
          type: integer
        prazo:
          type: string
        notas:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - titulo
...